pyproject.toml
[project]
name = "leadpulse"
version = "1.0.0"
description = "Automated Business Lead Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "sqlalchemy>=2.0.31",
    "jinja2>=3.1.4",
    "pydantic>=2.7.4",
    "python-multipart>=0.0.9",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "tldextract>=5.1.2",
    "dnspython>=2.6.1",
    "pandas>=2.2.2",
    "openpyxl>=3.1.5",
    "spacy>=3.7.5",
]
